Fix: handle texts with NUL char (#5).
The text in rare cases may contain a NUL character which is NOT
meant as a string-terminator. It broke down assertion that text has
the same size and char bboxes.

The char pointer with bytes was implicitly converted to str (Python2's
"bytes") with NUL termination. Thus the decoded text was not complete.

But since _text (a GooString) knows it's size we can convert that char
pointer to python bytes without NUL termination using the array slice syntax!